+++ VendorAndOrderTracker/Models/Order.cs
@@ -6,12 +6,12 @@ namespace VendorAndOrderTracker.Models
   {
     public string Title { get; set; }
     public string Description { get; set; }
-    public int Price { get; set; }
+    public decimal Price { get; set; }
     public int OrderId { get; }
     public string Date { get; set; }
 
     private static List<Order> _instances = new List<Order> {};
-    public Order (string title, string description, int price, string date)
+    public Order (string title, string description, decimal price, string date)
     {
       Title = title;
       Description = description;
